Separate in header and source files
